requirements.txt
opencv-python
pytesseract
python-dotenv 
python-telegram-bot
